requirements.txt
Django==2.0.7
django-filter==2.0.0
pytz==2018.5
